refactor: (#205) wus 폴더에 있던 유저 MSW 코드 밖으로 이동
