+++ frontend/src/mocks/handlers.ts
@@ -3,9 +3,9 @@ import { mockComment } from './comment';
 import { example } from './example/get';
 import { mockVoteResult } from './getVoteDetail';
 import { mockPost } from './post';
+import { mockUserInfo } from './userInfo';
 import { mockVote } from './vote';
 import { mockPostList } from './wus/post';
-import { mockUserInfo } from './wus/userInfo';
 
 export const handlers = [
   ...example,
